conversion: component.js to tsx
